🔨 refactor: qna page init
