--- front/src/components/Qna.js
@@ -1,7 +1,0 @@
-import React from "react";
-
-const Qna = ({ style }) => {
-  return <div style={style}></div>;
-};
-
-export default Qna;
